requirements.txt
azure-core==1.30.1
azure-functions==1.19.0
azure-identity==1.16.0
azure-search==1.0.0b2
azure-search-documents==11.6.0b3
langchain-openai==0.1.7
openai==1.30.1
python-dotenv==1.0.1